+++ test_program/CMakeLists.txt
@@ -18,7 +18,7 @@ set(LUA_LIBRARY_DIR "${LUA_DIR}")
 
 # Add the executable
 add_executable(test
-        test.cpp
+        main.cpp
 )
 
 # Include directories for lua headers
