+++ test_program/lua/README.md
@@ -1,3 +1,5 @@
 # Note
 
 These files are automatically generated by the `convert_lua.sh` and `build.sh` scripts.
+
+The source files are in `./lua_src`
